New Symfony 4 bin/composer directory structure (#9)
* New Symfony 4 bin/composer directory structure

Hi thanks for this helpful bundle. The update script anticipates the need to use 'bin' rather than 'app' but the method useNewDirectoryStructure is looking for symfony-var-dir option.  I added the symfony-bin-dir directory as an explicit option and the useNewDirectoryStructure method to look for that.  I think this will help keep the script working with projects independent of 'symfony-var-dir' .  Thanks!

* Update ScriptHandler.php

* Update ScriptHandler.php

* Update README.md

Took a deeper dive on this and it turns out we can just call the Command directly.  Updated the docs, thank you.

* Update README.md